% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_per_reader_nest.R
\name{AUC_cov_2reader_nest}
\alias{AUC_cov_2reader_nest}
\title{Function for calculating 2 reader AUC covariance}
\usage{
AUC_cov_2reader_nest(success_score)
}
\arguments{
\item{success_score}{The success score for nested data}
}
\value{
the covariance between two readers' AUC
}
\description{
Function for calculating 2 reader AUC covariance
}
