% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_per_reader_nest.R
\name{cov_m8_f2}
\alias{cov_m8_f2}
\title{covariance 8th moment middle calculation part two}
\usage{
cov_m8_f2(m)
}
\arguments{
\item{m}{input matrix}
}
\value{
the middle values for calculating covariance 8th moment
}
\description{
covariance 8th moment middle calculation part two
}
