% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NancySimulation.R
\name{data_MRMC}
\alias{data_MRMC}
\title{Simulation function}
\usage{
data_MRMC(sim.config)
}
\arguments{
\item{sim.config}{list contains following parameters:
I  num The number of patients.
k  num The number of ROIs in each patient.
R num The number of readers.
correlation_t num The correlation for simulating truth label.
potential_correlation_s num The correlation for simulating reading scores.
AUC_all num The theoretical AUC values.
sameclustersize boolean The binary variable to decide whether we have same number of ROIs in each patient.
rho num The scale parameter that infulence the covariance matrix in multivariate normal distribution.
fix_design boolean Binary variable to decide whether fix the truth label in simulation.
stream num The integer control the random number generator.}
}
\value{
A list and the only element in the list is the simulated data with following columns:
"clusterID","unitID","reader1",...,"truth"
}
\description{
Simulation function
}
