% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_per_reader_nest.R
\name{delete_diag}
\alias{delete_diag}
\title{Delete diagonal term function}
\usage{
delete_diag(m)
}
\arguments{
\item{m}{the input matrix for deleting diagonal term}
}
\value{
diagonal term removed matrix
}
\description{
Delete diagonal term function
}
