% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NancySimulation.R
\docType{data}
\name{expected_data}
\alias{expected_data}
\title{The test demo data to be included in my package}
\description{
The test demo data to be included in my package
}
\author{
Hongfei Du \email{hongfei@gwu.edu}
}
\keyword{data}
