% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_per_reader_nest.R
\name{m11_f}
\alias{m11_f}
\title{function for calculating the 11th moment}
\usage{
m11_f(m)
}
\arguments{
\item{m}{input matrix}
}
\value{
the 11th moment
}
\description{
function for calculating the 11th moment
}
