% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_per_reader_nest.R
\name{m8_f}
\alias{m8_f}
\title{function for calculating the 8th moment}
\usage{
m8_f(m)
}
\arguments{
\item{m}{input matrix}
}
\value{
the 8th moment
}
\description{
function for calculating the 8th moment
}
