% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_per_reader_nest.R
\name{success_score}
\alias{success_score}
\title{Calculate the success score}
\usage{
success_score(data)
}
\arguments{
\item{data}{the nested MRMC data}
}
\value{
The success score and number of ROIs in each case
}
\description{
Calculate the success score
}
