% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_coef}
\alias{var_coef}
\title{Calculate the each moments coefficient in variance}
\usage{
var_coef(numROI)
}
\arguments{
\item{numROI}{number of ROIs in each case}
}
\value{
all the coefficients
}
\description{
Calculate the each moments coefficient in variance
}
